fixed issue where stats wouldnt be saved. because of dontdestroyonload singleton severing inspector unity events.

+++ Assets/Game/Scripts/CustomStatsManager.cs
@@ -44,7 +44,7 @@ namespace Game.Scripts {
         public int playerHealth;
         public float playerSpeed;
         public float playerAttackSpeed;
-        public int enemyHealthMult;
+        public float enemyHealthMult;
         public float enemyAttackSpeedMult;
         
         public string[] ToArray() {
